+++ src/types.h
@@ -16,7 +16,7 @@ using namespace Chess;
 
 #define IS_DEBUG true
 
-typedef uint64_t Time;
+typedef double Time;
 
 static inline bool is_capture(Board &board, Move move) { return (board.pieceAtB(to(move)) != None); }
